+++ packages/sync-core/src/lib/TLSocketRoom.ts
@@ -1,6 +1,6 @@
 import type { StoreSchema, UnknownRecord } from '@tldraw/store'
 import { TLStoreSnapshot, createTLSchema } from '@tldraw/tlschema'
-import { objectMapValues, structuredClone, uniqueId } from '@tldraw/utils'
+import { objectMapValues, structuredClone } from '@tldraw/utils'
 import { RoomSessionState } from './RoomSession'
 import { ServerSocketAdapter, WebSocketMinimal } from './ServerSocketAdapter'
 import { TLSyncErrorCloseEventReason } from './TLSyncClient'
@@ -314,41 +314,6 @@ export class TLSocketRoom<R extends UnknownRecord = UnknownRecord, SessionMeta =
 		// replace room with new one and kick out all the clients
 		this.room = newRoom
 		oldRoom.close()
-	}
-
-	/** @internal */
-	async associateFileAssets(fileId: string, bucket: any) {
-		const assetsToUpdate: { objectName: string; fileId: string }[] = []
-		await this.updateStore(async (store) => {
-			const records = store.getAll()
-
-			for (const record of records) {
-				if (record.typeName !== 'asset') continue
-				const asset = record as any
-				const meta = asset.meta
-
-				if (meta?.fileId === fileId) continue
-				const src = asset.props.src
-				if (!src) continue
-				const objectName = src.split('/').pop()
-				if (!objectName) continue
-				const currentAsset = await bucket.get(objectName)
-				if (!currentAsset) continue
-
-				const split = objectName.split('-')
-				const fileType = split.length > 1 ? split.pop() : null
-				const id = uniqueId()
-				const newObjectName = fileType ? `${id}-${fileType}` : id
-				await bucket.put(newObjectName, currentAsset.body, {
-					httpMetadata: currentAsset.httpMetadata,
-				})
-				asset.props.src = asset.props.src.replace(objectName, newObjectName)
-				asset.meta.fileId = fileId
-				store.put(asset)
-				assetsToUpdate.push({ objectName: newObjectName, fileId })
-			}
-		})
-		return assetsToUpdate
 	}
 
 	/**
